Ai Agent/requirements.txt
phidata>=2.0.0
pymongo>=4.5.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
email-validator>=2.0.0
google-generativeai>=0.3.0
apscheduler>=3.10.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
